Renamed Libgen plugin
